requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
pydantic==2.12.4
pydantic-settings==2.12.0
sqlalchemy==2.0.44
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pytest-asyncio>=0.21.0
pytest-cov==7.0.0
requests==2.31.0
jinja2==3.1.5
a2wsgi==1.10.7
psycopg2-binary==2.9.9
email-validator==2.1.0.post1
argon2-cffi==23.1.0
httpx>=0.28.1
tenacity==9.0.0
alembic>=1.14.0
resend>=2.23.0
pyotp>=2.9.0
prometheus-client>=0.21.0
opentelemetry-distro==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
python-json-logger==2.0.7
networkx>=3.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.0
